Fix: add failsafe if no travis build is present
The new fixes navigate through pages of statuses to check for CI contexts.

If no CI status url was given, treat the PR as normal (a passed build).
If no CI context was found in any statuses page, treat the PR as normal.
If a CI context was found, return true iff the state was 'success', else return false.